refactor(ESLint): rename partial pkg config files
