+++ packages-proprietary/react-icons/eslint.config-partial.mjs
@@ -1,4 +1,4 @@
-import { reactPreset } from '../../packages/react/eslint.react.config.mjs'
+import { reactPreset } from '../../packages/react/eslint.config-partial.mjs'
 
 /* Imported and spread by the root `eslint.config.mjs`. */
 export default [
