+++ packages/react/eslint.config-partial.mjs
